+++ src/index.js
@@ -35,6 +35,6 @@ const configureSession = (paymentSession, config) => {
   return sessionInstance.initialize()
 }
 
-export { HostedSessionError, HostedSessionValidationError } from './hosted-session'
+export { HostedSessionError } from './hosted-session'
 
 export default getHostedSession